# Optimize line drawing by avoiding per-tile work