--- a/js/render/draw_line.js
+++ b/js/render/draw_line.js
@@ -24,72 +24,84 @@ module.exports = function drawLine(painter, sourceCache, layer, coords) {
     // don't draw zero-width lines
     if (layer.paint['line-width'] <= 0) return;
 
+    const programId =
+        layer.paint['line-dasharray'] ? 'lineSDF' :
+        layer.paint['line-pattern'] ? 'linePattern' : 'line';
+
     for (let k = 0; k < coords.length; k++) {
-        drawLineTile(painter, sourceCache, layer, coords[k]);
+        const tile = sourceCache.getTile(coords[k]);
+        const bucket = tile.getBucket(layer);
+        if (!bucket) continue;
+
+        const layerData = bucket.buffers.layerData[layer.id];
+        const prevProgram = painter.currentProgram;
+        const program = painter.useProgram(programId, layerData.programConfiguration);
+        const programChanged = k === 0 || program !== prevProgram;
+
+        if (programChanged) {
+            layerData.programConfiguration.setUniforms(painter.gl, program, layer, {zoom: painter.transform.zoom});
+        }
+        drawLineTile(program, painter, tile, bucket.buffers, layer, coords[k], layerData, programChanged);
     }
 };
 
-function drawLineTile(painter, sourceCache, layer, coord) {
-    const tile = sourceCache.getTile(coord);
-    const bucket = tile.getBucket(layer);
-    if (!bucket) return;
-
-    const buffers = bucket.buffers;
-    const layerData = buffers.layerData[layer.id];
+function drawLineTile(program, painter, tile, buffers, layer, coord, layerData, programChanged) {
     const gl = painter.gl;
-
     const dasharray = layer.paint['line-dasharray'];
     const image = layer.paint['line-pattern'];
 
-    const programConfiguration = layerData.programConfiguration;
-    const program = painter.useProgram(dasharray ? 'lineSDF' : image ? 'linePattern' : 'line', programConfiguration);
-    programConfiguration.setUniforms(gl, program, layer, {zoom: painter.transform.zoom});
-
-    if (!image) {
-        gl.uniform4fv(program.u_color, layer.paint['line-color']);
-    }
-
     let posA, posB, imagePosA, imagePosB;
+
     if (dasharray) {
         posA = painter.lineAtlas.getDash(dasharray.from, layer.layout['line-cap'] === 'round');
         posB = painter.lineAtlas.getDash(dasharray.to, layer.layout['line-cap'] === 'round');
 
-        gl.uniform1i(program.u_image, 0);
-        gl.activeTexture(gl.TEXTURE0);
-        painter.lineAtlas.bind(gl);
-
-        gl.uniform1f(program.u_tex_y_a, posA.y);
-        gl.uniform1f(program.u_tex_y_b, posB.y);
-        gl.uniform1f(program.u_mix, dasharray.t);
-
     } else if (image) {
         imagePosA = painter.spriteAtlas.getPosition(image.from, true);
         imagePosB = painter.spriteAtlas.getPosition(image.to, true);
-        if (!imagePosA || !imagePosB) return;
-
-        gl.uniform1i(program.u_image, 0);
-        gl.activeTexture(gl.TEXTURE0);
-        painter.spriteAtlas.bind(gl, true);
-
-        gl.uniform2fv(program.u_pattern_tl_a, imagePosA.tl);
-        gl.uniform2fv(program.u_pattern_br_a, imagePosA.br);
-        gl.uniform2fv(program.u_pattern_tl_b, imagePosB.tl);
-        gl.uniform2fv(program.u_pattern_br_b, imagePosB.br);
-        gl.uniform1f(program.u_fade, image.t);
     }
 
-    // the distance over which the line edge fades out.
-    // Retina devices need a smaller distance to avoid aliasing.
-    const antialiasing = 1 / browser.devicePixelRatio;
+    if (programChanged) {
+        if (!image) {
+            gl.uniform4fv(program.u_color, layer.paint['line-color']);
+        }
 
-    gl.uniform1f(program.u_linewidth, layer.paint['line-width'] / 2);
-    gl.uniform1f(program.u_gapwidth, layer.paint['line-gap-width'] / 2);
-    gl.uniform1f(program.u_antialiasing, antialiasing / 2);
-    gl.uniform1f(program.u_blur, layer.paint['line-blur'] + antialiasing);
-    gl.uniform1f(program.u_opacity, layer.paint['line-opacity']);
-    gl.uniformMatrix2fv(program.u_antialiasingmatrix, false, painter.transform.lineAntialiasingMatrix);
-    gl.uniform1f(program.u_offset, -layer.paint['line-offset']);
-    gl.uniform1f(program.u_extra, painter.transform.lineStretch);
+        if (dasharray) {
+            gl.uniform1i(program.u_image, 0);
+            gl.activeTexture(gl.TEXTURE0);
+            painter.lineAtlas.bind(gl);
+
+            gl.uniform1f(program.u_tex_y_a, posA.y);
+            gl.uniform1f(program.u_tex_y_b, posB.y);
+            gl.uniform1f(program.u_mix, dasharray.t);
+
+        } else if (image) {
+            if (!imagePosA || !imagePosB) return;
+
+            gl.uniform1i(program.u_image, 0);
+            gl.activeTexture(gl.TEXTURE0);
+            painter.spriteAtlas.bind(gl, true);
+
+            gl.uniform2fv(program.u_pattern_tl_a, imagePosA.tl);
+            gl.uniform2fv(program.u_pattern_br_a, imagePosA.br);
+            gl.uniform2fv(program.u_pattern_tl_b, imagePosB.tl);
+            gl.uniform2fv(program.u_pattern_br_b, imagePosB.br);
+            gl.uniform1f(program.u_fade, image.t);
+        }
+
+        // the distance over which the line edge fades out.
+        // Retina devices need a smaller distance to avoid aliasing.
+        const antialiasing = 1 / browser.devicePixelRatio;
+
+        gl.uniform1f(program.u_linewidth, layer.paint['line-width'] / 2);
+        gl.uniform1f(program.u_gapwidth, layer.paint['line-gap-width'] / 2);
+        gl.uniform1f(program.u_antialiasing, antialiasing / 2);
+        gl.uniform1f(program.u_blur, layer.paint['line-blur'] + antialiasing);
+        gl.uniform1f(program.u_opacity, layer.paint['line-opacity']);
+        gl.uniformMatrix2fv(program.u_antialiasingmatrix, false, painter.transform.lineAntialiasingMatrix);
+        gl.uniform1f(program.u_offset, -layer.paint['line-offset']);
+        gl.uniform1f(program.u_extra, painter.transform.lineStretch);
+    }
 
     painter.enableTileClippingMask(coord);