This is a slightly messy attempt at winning back some of the performance that regressed with #3342 back, until we solve the challenge of ensuring all tiles are always rendered with the same style.

The idea is to simply check whether the program changed after `useProgram` call, and not do a big portion of the drawing work if it didn't.

On my [3342 benchmark](http://jsfiddle.net/Mourner/2bzyLLp5/4/), results are as follows:

- v0.19: 8.3ms
- v0.25: 17.2ms
- master: 15.3ms (looks like some of your recent changes saved us some ms @jfirebaugh!)
- this PR: ~~12.2ms~~ 11.7ms

👀 @lucaswoj @jfirebaugh