@jfirebaugh 2016-10-28
I'd prefer to follow the optimization that native uses: each program object records the value which was last set for each uniform, and avoids calling `glUniform*` when the value isn't changing. This optimization benefits every layer type as well as the `k === 0` case here (you can reuse the uniform values set by a previous layer, if they're not changing).