@mourner 2016-10-28
@jfirebaugh yes, but that sounds like a more involved refactor (#145). Meanwhile this PR is an easy short-term stopgap that we can revert at any time.