@mourner 2016-10-28
@jfirebaugh added another commit that removes redundant state changes for pattern/dash-related line drawing code by tracking if the tile zoom changed between tiles. Also cleaned it up a bit. This brings the benchmark result a little further down, from 12.2 to 11.7.