@mourner 2016-10-28
`frame-duration` benchmark result isn't as significant because of the averaging I mentioned in #3342, but still noticeable:

benchmark | master | optimize-line-draw
--- | --- | ---
**frame-duration** | 9.6 ms, 5% > 16ms  | 8.9 ms, 1% > 16ms